+++ index.js
@@ -1,4 +1,4 @@
-// version: v1.0.28 | build: 2026-07-25T20:55:20Z
+// version: v1.0.30 | build: 2026-07-25T21:26:08Z
 let products = [], orders = [], customers = [];
 let settings = {};
 let cQty = {};